Add random block selection for context and target blocks
- Create constants.rs module with configurable block counts
- Generate unique random numbers for context and target block selection
- Update random_whole_number() to use usize for better array indexing
- Add logic to ensure no duplicate block numbers are selected
- Fix file path comments across multiple modules

+++ implementation/src/sundry.rs
@@ -1,11 +1,11 @@
 /*
- * implementation/sundry.rs
+ * implementation/src/sundry.rs
  * Q@khaa.pk
  */
 
 use rand::Rng;
 
-pub fn random_whole_number(min: u32, max: u32) -> u32 {
+pub fn random_whole_number(min: usize, max: usize) -> usize {
 
     let mut rng = rand::thread_rng();
     rng.gen_range(min..=max)
